pyproject.toml
[tool.poetry]
name = "asymetry"
version = "0.1.1-alpha"
description = "LLM observability SDK for OpenAI and Anthropic"
authors = ["Asymetry Team"]
license = "MIT"
readme = "README.md"
homepage = "https://asymetry.co"
repository = "https://github.com/asymetry/asymetry-sdk"
documentation = "https://docs.asymetry.co"
keywords = ["llm", "observability", "openai", "monitoring", "tracing", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
packages = [{include = "asymetry"}]

[tool.poetry.dependencies]
python = "^3.13"
openai = "^1.0.0"
anthropic = "0.72.0"
httpx = "^0.27.0"
tiktoken = "^0.8.0"
python-dotenv = "^1.0.0"
opentelemetry-api = "1.20.0"
opentelemetry-sdk = "1.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"